feat(pages): add `settings` page and nested page
- settings page will be a list item, can be navigate to nested settings
- add `language and region` item/page by current
- add `about` item/page by current
