+++ src/pages/LanguageSettingsPage.tsx
@@ -1,0 +1,21 @@
+import { IonBackButton, IonButtons, IonContent, IonHeader, IonPage, IonText, IonTitle, IonToolbar } from '@ionic/react';
+
+function LanguageSettingsPage() {
+  return (
+    <IonPage>
+      <IonHeader>
+        <IonToolbar>
+          <IonButtons slot="start">
+            <IonBackButton></IonBackButton>
+          </IonButtons>
+          <IonTitle>Language & Region</IonTitle>
+        </IonToolbar>
+      </IonHeader>
+      <IonContent fullscreen>
+        <IonText>Language</IonText>
+      </IonContent>
+    </IonPage>
+  );
+}
+
+export { LanguageSettingsPage };
